# Replace synchronization in Gauge with the use of volatile