@brian-brazil 2016-04-06
I don't believe this is correct as it introduces races in `inc` and `dec`. It'll also be slower for these, as there's now a single value that has to be keep consistent across CPU caches rather than at least one per active core.