@pkubowicz 2016-04-06
How can it introduce races if the Java language specification (see the link above) guarantees this to be safe?

As for the performance - see the benchmarks. Do you have any facts against them? The cost of synchronization under contention is severe, much more important than any caching. This class won't be fast if you use an exclusive lock and this is how it's done now.