@beorn7 2016-04-06
@pkubowicz You change does introduce races in inc() and dec(). the `+=` and `-=` operator is not atomic, even if a write to the volatile double is. Those operators first read the variable, then increment or decrement the value, and then store the value back into the variable. In between the (atomic) read and write, another thread may call inc() or dec(), and then you have a race.