@brian-brazil 2016-04-06
It guarantees the read to be safe and the write to be safe - it doesn't guarantee that read, increment and then write is safe.

You've sped up set (which we know is slow) by 3.5x but slowed down inc and dec by a factor of 4. If we just had to offer set then this might be an option, but we must offer both sets of functions.