@pkubowicz 2016-04-06
Ok, I admit my mistake. Guava has an AtomicDouble. Maybe it has performance similar to `volatile`.

> You've sped up set (which we know is slow) by 3.5x but slowed down inc and dec by a factor of 4. If we just had to offer set then this might be an option, but we must offer both sets of functions.

I have increased predictability: operations will take similar time. I think it's weird when `set` is more than 10 times slower than `inc` and this is how it's now.

I think for Gauge `set` is an important operation, if not the most important. If you don't need `set`, you use Counter. Plus, you don't count the synchronization penalty for each `get`. If many threads set a gauge in the same time, collecting metrics will block, and this is not a good behaviour.

Coming back to synchronization, I think the current code has to be incorrect, because you access the DoubleAdder both in a synchronized an in a unsynchronized way. Imaging calls:
```
// gauge is 500
gauge.inc(5); // thread 1
gauge.set(1000); // thread2
```
current sychronization allows the following sequence of JVM instructions:
```
lock(this) // from set()
value.reset(); // from set()
value.add(5); // from inc() - does not synchronize, so it's allowed here
value.add(1000); // from set()
unlock(this) // from set()
// gauge is 1005
```