@brian-brazil 2016-04-06
> I think for Gauge set is an important operation, if not the most important.

Inc and dec are actually more important, as they allow you to update the gauge as you go rather than having to maintain a gauge in a variable yourself and do some form of callback.

> If you don't need set, you use Counter

This is a misconception that other metrics systems have spread. Counters can't have a dec method, otherwise they're not counters. 

> I think the current code has to be incorrect, because you access the DoubleAdder both in a synchronized an in a unsynchronized way.

If you've mixing inc/dec and set, then it's undefined in what order they'll resolve so this is fine and your example produces a valid output.

In practice for a given gauge you'll either be using one or the other.